Trade: move router locally
