+++ src/apps/chat/trade/TradeModal.tsx
@@ -4,8 +4,8 @@ import { Divider } from '@mui/joy';
 
 import { GoodModal } from '~/common/components/GoodModal';
 
-import { ImportConfig, ImportConversations } from './ImportChats';
 import { ExportConfig, ExportChats } from './ExportChats';
+import { ImportConfig, ImportConversations } from './ImportChats';
 
 export type TradeConfig = ImportConfig | ExportConfig;
