+++ src/apps/chat/trade/server/publish.pastegg.ts
